Create tests for adding and removing tasks
